Add experimental linux support
It is known to work for Lubuntu 16.10.
Modify Gruntfile to include a Linux build. Also add instructions to
README.md to guide users on how to install Cumulus on Linux, and more
specifically on Ubuntu flavors. To help users integrate the app into
their desktop environment, a Cumulus.desktop file has been added with
all the needed values already provided.